requirements.txt
python-dotenv==1.0.1
requests==2.32.3
fastapi==0.115.12
uvicorn==0.30.3
jinja2==3.1.4
python-multipart==0.0.9
python-docx==1.1.2
starlette==0.46.2
sse-starlette==2.3.6 